autocomplete items on checklist page change instead of waiting for function call from timer
